2k26/02/RUL/code/requirements.txt
-e .
ensure
Flask
Flask-Cors
dvc
dvc[gdrive]
gunicorn
joblib
numpy
pandas
python-box
PyYAML
scikit-learn
setuptools
tensorflow